pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.13.0",
    "litellm>=1.66.3",
    "ollama>=0.5.3",
    "openai>=1.102.0",
    "yfinance>=0.2.65",
    "google-api-python-client>=2.88.0",
    "google-auth-oauthlib>=0.8.0",
]
